refactor(arch): move move_metric.py to bloqade.lanes.arch.metrics (stage 7 of #569)
Seventh stage of the multi-stage package merge tracked in #569.
Renames bloqade.lanes.layout.move_metric to bloqade.lanes.arch.metrics
in one move — move-metric computation builds on top of the runtime
arch surface, so it lives alongside it. The shorter ``metrics``
filename also drops the now-redundant ``move_`` prefix.

Old path bloqade.lanes.layout.move_metric remains as a thin
re-export shim. Removed in the final cleanup stage.

Refs #569.
